refactor: revise Vistiation page
