+++ components/templates/Visitation/Visitation.tsx
@@ -1,23 +1,26 @@
+import type { NextPage } from 'next';
 import { useRef, useMemo } from 'react';
 
 import { Visitation } from 'api/visitations';
 import { useTab } from 'util/hooks/useTab';
 
+import TemplateWrapper from 'components/organisms/TemplateWrapper';
 import ContentRender from 'components/molecules/ContentRender';
 import TabNav from 'components/molecules/TabNav';
 import SectionWrapper from 'components/molecules/SectionWrapper';
 
-import styles from './Content.module.scss';
+import { appPath } from 'util/routing.config';
 
+import styles from './Visitation.module.scss';
+
+interface IProps {
+  visitations: Visitation[];
+}
 interface VisitationMap {
   [name: string]: Visitation;
 }
 
-interface ContentProps {
-  visitations: Visitation[];
-}
-
-const Content = ({ visitations }: ContentProps) => {
+const Content = ({ visitations }: IProps) => {
   const elRef = useRef<HTMLDivElement>(null);
   const tabs = useMemo(
     () =>
@@ -68,4 +71,20 @@ const Content = ({ visitations }: ContentProps) => {
   );
 };
 
-export default Content;
+const Visitation: NextPage<IProps> = ({ visitations }: IProps) => {
+  return (
+    <TemplateWrapper
+      primaryTitle="企業參訪"
+      subTitle="參訪宗旨"
+      description="透過不同產業的企業參訪，了解各產業的趨勢，並藉此深入接觸企業文化、職場環境、公司發展與未來規劃。讓學員能更貼近職場，也更能增進對未來職涯探索的規劃。"
+      desktopBackgroundImage="/images/faq/pics/banner--desktop.png"
+      mobileBackgroundImage="/images/faq/pics/banner--mobile.png"
+      nextSectionTitle="計畫介紹"
+      nextSectionPath={appPath.plan}
+    >
+      <Content visitations={visitations} />
+    </TemplateWrapper>
+  );
+};
+
+export default Visitation;
